SQL Normalizer
Welcome to the SQL Normalizer project! This Python-based tool is designed to assist in the normalization process of relational databases. By providing a list of functional dependencies and attributes, the SQL Normalizer helps identify the primary key of the relation and facilitates normalization to either the Third Normal Form (3.NF) or Boyce-Codd Normal Form (BCNF).

Key Features
Functional Dependency Analysis: The SQL Normalizer analyzes the provided functional dependencies to determine the primary key of the relation.
Normalization to 3.NF or BCNF: Based on the identified primary key, the tool normalizes the relation to either the Third Normal Form (3.NF) or Boyce-Codd Normal Form (BCNF), ensuring data integrity and eliminating redundancies.
Python-Powered: Built using Python, the SQL Normalizer is a versatile and efficient solution for database normalization.